+++ R/as_draws_rvars.R
@@ -226,6 +226,20 @@ as_draws_rvars.mcmc.list <- function(x, ...) {
   x <- conform_rvar_ndraws_nchains(x)
 
   class(x) <- class_draws_rvars()
+
+  # move the .log_weight column into the log_weights attribute of each rvar,
+  # but only if there is no conflict between any existing weights on the rvars
+  if (".log_weight" %in% names(x)) {
+    existing_weights <- log_weights(x[[1]])
+    .log_weight <- as.vector(draws_of(x$.log_weight))
+    if (is.null(existing_weights)) {
+      x$.log_weight <- NULL
+      x <- weight_draws(x, .log_weight, log = TRUE)
+    } else {
+      weights2_common(existing_weights, .log_weight)
+    }
+  }
+
   x
 }
 
@@ -273,4 +287,14 @@ empty_draws_rvars <- function(variables = character(0), nchains = 0) {
   out <- named_list(variables, rvar())
   class(out) <- class_draws_rvars()
   out
+}
+
+# when converting draws_rvars to other formats, we must promote log weights
+# to be a variable before doing the conversion
+promote_rvar_weights_to_variable <- function(x) {
+  .log_weights <- log_weights(x)
+  if (!is.null(.log_weights)) {
+    x$.log_weight <- rvar(log_weights(x))
+  }
+  x
 }